separate custom gradle scripts into own files
